chore: Use IOperatingSystem.Unix singletone
