+++ src/Testcontainers/Configurations/Windows.cs
@@ -49,6 +49,12 @@ namespace DotNet.Testcontainers.Configurations
       DockerEndpointAuthConfig = dockerEndpointAuthConfig;
     }
 
+    /// <summary>
+    /// Gets the <see cref="IOperatingSystem" /> instance.
+    /// </summary>
+    public static IOperatingSystem Instance { get; }
+      = new Windows(dockerEndpointAuthConfig: null);
+
     /// <inheritdoc />
     public IDockerEndpointAuthenticationConfiguration DockerEndpointAuthConfig { get; }
